requirements.txt
openai-whisper
opencv-python
moviepy
torch
numpy
tqdm